A static assert and comment about the session key
I'm feeding the mceliece kem'd key (which is hashed, and 32 bytes)
directly into crypto_secretbox (which wants a 32 byte key). This seems
ok to me, but I'm not a cryptographer and I would think that, wouldn't
I...

In any case, if this static_assert fails, you are in for a bad time.