Tweaks to the "nomalloc" functions, and a test
I think it's better to have the mcleece_cbox_seal message buffer sized
to the larger number -- that's the thing that explodes if you get it
wrong. We can do the math inside the function to trim things down to
size.

Also, I think I've been passing too many references...